+++ packages/widgetbook/lib/src/navigation/icons/icons.dart
@@ -1,3 +1,0 @@
-export 'component_icon.dart';
-export 'expander_icon.dart';
-export 'story_icon.dart';