+++ sandbox/lib/widgetbook.dart
@@ -15,15 +15,6 @@ class WidgetbookApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Widgetbook.material(
-      directories: [
-        WidgetbookLeafComponent(
-          name: 'Container',
-          useCase: WidgetbookUseCase.child(
-            name: 'Default',
-            child: const FlutterLogo(),
-          ),
-        ),
-      ],
       components: components,
       integrations: [
         WidgetbookCloudIntegration(),